Eliminate mediating class
